Move the UserInfo interface to the identity module
